updated fixes

+++ resources/views/users/index.blade.php
@@ -7,9 +7,11 @@
         <div class="pull-left">
             <h2>Users Management</h2>
         </div>
+        @can('user-create')
         <div class="pull-right">
             <a class="btn btn-success" href="{{ route('users.create') }}"> Create New User</a>
         </div>
+        @endcan
     </div>
 </div>
 
@@ -43,10 +45,14 @@
     </td>
     <td>
        <a class="btn btn-info" href="{{ route('users.show',$user->id) }}">Show</a>
+       @can('user-edit')
        <a class="btn btn-primary" href="{{ route('users.edit',$user->id) }}">Edit</a>
+       @endcan
+       @can('user-delete')
         {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
             {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
-        {!! Form::close() !!}
+          {!! Form::close() !!}
+        @endcan
     </td>
   </tr>
  @endforeach
